Batch payment rewards via view header
